test: refactor tests
Split the big test/test.js file into 4 smaller files
